Review changes

+++ python/cudf/cudf/core/series.py
@@ -3778,11 +3778,7 @@ class Series(Frame, Serializable):
             data = _format_stats_values(data)
 
             return Series(
-                data=data,
-                index=names,
-                nan_as_null=False,
-                name=self.name,
-                dtype="float",
+                data=data, index=names, nan_as_null=False, name=self.name,
             )
 
         def describe_categorical(self):
